@@ -1,52 +1,57 @@
 cmake_minimum_required(VERSION 3.3)
-set(CMAKE_CXX_STANDARD 14)  # TODO: C++17 when gcc 7 & clang 4 debain testing release; cmake_minimum_required -> 3.8
-project(fancon)
+project(fancon LANGUAGES CXX)
 set(PACKAGE_AUTHOR "Hayden Briese <haydenbriese@gmail.com>")
 
-## Additional cmake modules provided in ./cmake/
-list(APPEND CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake)
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-## Optimization flag - lvl 2
-set(OPTIMIZATION_FLAG "-O2")
+## Clang must be used because of linking with Boost Threads
+set(CMAKE_C_COMPILER "clang")
+set(CMAKE_CXX_COMPILER "clang++")
+
+## Additional cmake modules provided in ./cmake/
+set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${PROJECT_SOURCE_DIR}/cmake")
+
+## Default to Release if not set
+if (NOT CMAKE_BUILD_TYPE)
+    set(CMAKE_BUILD_TYPE Release)
+endif ()
+
+## Release flags
+set(OPTIMIZATION_FLAG "-O3")
+if (CMAKE_BUILD_TYPE STREQUAL "Release")
+    set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} ${OPTIMIZATION_FLAG}")
+endif ()
 
 ## Debug flags
 option(OPTIMIZE_DEBUG "Optimize debug build - useful for profiling" OFF)
 if (CMAKE_BUILD_TYPE STREQUAL "Debug")
-    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Wall -Wextra -Wpedantic -Werror")
+    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Wall -Wextra -Wpedantic")
     if (OPTIMIZE_DEBUG)
         set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} ${OPTIMIZATION_FLAG}")
     endif ()
 endif ()
 
-## Release flags
-if (CMAKE_BUILD_TYPE STREQUAL "Release")
-    set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} ${OPTIMIZATION_FLAG}")
-endif ()
-
+set(SRC "${PROJECT_SOURCE_DIR}/src")
+set(PSRC "${SRC}/proto")
 include_directories(src)
-set(SOURCE_FILES src/main.cpp src/main.hpp src/Util.cpp src/Util.hpp src/Logging.hpp src/Logging.cpp
-        src/UID.cpp src/UID.hpp src/Devices.cpp src/Devices.hpp src/Config.cpp src/Config.hpp
-        src/Fan.cpp src/Fan.hpp src/FanInterface.cpp src/FanInterface.hpp src/SensorInterface.cpp src/SensorInterface.hpp
-        src/NvidiaDevices.cpp src/NvidiaDevices.hpp src/NvidiaUtil.cpp src/NvidiaUtil.hpp src/Controller.cpp src/Controller.hpp)
+set(SOURCE_FILES ${SRC}/main.cpp ${SRC}/main.hpp
+        ${SRC}/Util.cpp ${SRC}/Util.hpp
+        ${SRC}/Logging.hpp ${SRC}/Logging.cpp
+        ${PSRC}/DevicesSpec.pb.cc ${PSRC}/DevicesSpec.pb.h
+        ${PSRC}/DevicesSpec.grpc.pb.cc ${PSRC}/DevicesSpec.grpc.pb.h
+        ${SRC}/Controller.cpp ${SRC}/Controller.hpp
+        ${SRC}/Devices.cpp ${SRC}/Devices.hpp
+        ${SRC}/FanInterface.cpp ${SRC}/FanInterface.hpp
+        ${SRC}/SensorInterface.cpp ${SRC}/SensorInterface.hpp
+        ${SRC}/FanSysfs.cpp ${SRC}/FanSysfs.hpp
+        ${SRC}/SensorSysfs.cpp ${SRC}/SensorSysfs.hpp
+        ${SRC}/NvidiaUtil.cpp ${SRC}/NvidiaUtil.hpp
+        ${SRC}/NvidiaDevices.cpp ${SRC}/NvidiaDevices.hpp
+        ${SRC}/NvidiaNvml.cpp ${SRC}/NvidiaNvml.hpp
+        ${SRC}/DellSMM.cpp ${SRC}/DellSMM.hpp
+        ${SRC}/FanDell.cpp ${SRC}/FanDell.hpp)
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
-
-# Set system paths
-if (NOT DEFINED CMAKE_INSTALL_SYSCONFDIR)
-    set(CMAKE_INSTALL_SYSCONFDIR "/etc")
-endif ()
-add_definitions(-DFANCON_SYSCONFDIR="${CMAKE_INSTALL_SYSCONFDIR}")
-
-if (NOT DEFINED CMAKE_INSTALL_LOCALSTATEDIR)
-    set(CMAKE_INSTALL_LOCALSTATEDIR "/var")
-endif ()
-add_definitions(-DFANCON_LOCALSTATEDIR="${CMAKE_INSTALL_LOCALSTATEDIR}")
-
-## libstdc++ static linking
-option(STATIC_LIBSTDC++ "Statically link libstdc++" OFF)
-if (STATIC_LIBSTDC++)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -static-libstdc++")
-    message("Statically linking libstdc++")
-endif ()
 
 ## Threads (pthread)
 find_package(Threads REQUIRED)
