@@ -1,62 +1,71 @@
 #ifndef FANCON_CONTROLLER_HPP
 #define FANCON_CONTROLLER_HPP
 
-#include "Devices.hpp"
-#include <algorithm> // find_if
-#include <chrono>
-#include <csignal>
-#include <sstream> // istringstream
+#include <algorithm>
+#include <boost/thread.hpp>
+#include <cmath>
+#include <future>
+#include <google/protobuf/text_format.h>
+#include <map>
+#include <sstream>
 #include <thread>
 
-using std::istringstream;
+#include "Devices.hpp"
+#include "proto/DevicesSpec.pb.h"
+
+using boost::thread;
+using fc::FanInterface;
 using std::find_if;
-using std::thread;
-using std::this_thread::sleep_for;
-using std::this_thread::sleep_until;
+using std::future;
+using std::istringstream;
 
-namespace fancon {
-struct MappedFan;
+namespace fc {
+extern bool dynamic;
+extern uint smoothing_intervals;
+extern uint top_stickiness_intervals;
+extern uint temp_averaging_intervals;
 
-using sensor_container_t = vector<unique_ptr<SensorInterface>>;
-using fan_container_t = vector<MappedFan>;
-using time_point_t = chrono::time_point<chrono::steady_clock, milliseconds>;
-
-enum class ControllerState {
-  run, stop = SIGTERM, reload = SIGHUP
+enum class ControllerState : sig_atomic_t {
+  RUN,
+  STOP,
+  RELOAD
 } extern controller_state;
 
 class Controller {
 public:
-  explicit Controller(const string &configPath);
+  Controller(path conf_path_);
 
-  controller::Config conf;
+  path config_path;
+  milliseconds update_interval{1000};
+  Devices devices;
 
-  time_point_t sensors_wakeup, fans_wakeup;
+  void start();
+  void test(bool force = false, bool safely = false);
 
-  sensor_container_t sensors;
-  fan_container_t fans;
-
-  ControllerState run();
-
-  void updateSensors(vector<sensor_container_t::iterator> sensors);
-  void updateFans(vector<fan_container_t::iterator> fans);
-
-  void scheduler();
-
-  static void signalHandler(int sig);
-  static bool validConfigLine(const string &line);
+  static bool running();
+  static void stop();
+  static void reload();
+  static bool reloading();
 
 private:
-  void deferStart(time_point_t &timePoint);
-};
+  fs::file_time_type config_write_time;
 
-struct MappedFan {
-  MappedFan(unique_ptr<FanInterface> fan, const SensorInterface &sensor)
-      : fan(move(fan)), sensor(sensor) {}
+  void from(const fc_pb::Controller &c);
+  void from(const fc_pb::ControllerConfig &c);
+  void to(fc_pb::Controller &c);
+  void to(fc_pb::ControllerConfig &c);
 
-  unique_ptr<FanInterface> fan;
-  const SensorInterface &sensor;
+  void load_devices();
+  void to_file(path file, bool backup = true);
+
+  void update_config_write_time();
+  bool config_file_modified() const;
+  static void shutdown_threads(vector<thread> &threads);
+  string date_time_now() const;
+
+  static bool tests_running(const vector<shared_ptr<double>> &test_completion);
+  static double sum(const vector<shared_ptr<double>> &numbers);
 };
-}
+} // namespace fc
 
 #endif // FANCON_CONTROLLER_HPP
