@@ -1,122 +1,96 @@
 #ifndef FANCON_FANINTERFACE_HPP
 #define FANCON_FANINTERFACE_HPP
 
-#include "Config.hpp"
-#include "UID.hpp"
+#include <cmath>
+#include <regex>
+
+#include "SensorInterface.hpp"
 #include "Util.hpp"
-#include <chrono>
-#include <sstream> // stringstream
-#include <thread>
-#include <tuple>
-#include <type_traits> // is_same
+#include "proto/DevicesSpec.pb.h"
 
-using std::this_thread::sleep_for;
-using std::stringstream;
-using fancon::UID;
-using fancon::Util::read;
-using fancon::Util::write;
-using fancon::Util::getPath;
+using std::abs;
+using std::min;
+using std::next;
+using std::stoi;
 
-using namespace fancon::fan;
+using Pwm = int;
+using Rpm = int;
+using Percent = uint;
+using Rpm_to_Pwm_Map = std::map<Rpm, Pwm>;
+using Pwm_to_Rpm_Map = std::map<Pwm, Rpm>;
+using Temp_to_Rpm_Map = std::map<Temp, Rpm>;
 
-namespace chrono = std::chrono;
+namespace fc {
+extern bool dynamic;
+extern uint smoothing_intervals;
+extern uint top_stickiness_intervals;
+enum class ControllerState;
+extern ControllerState controller_state;
 
-namespace fancon {
-struct FanCharacteristics;
-enum class FailState { null, control, pwm_write, rpm_read_accuracy };
-
-/// FanCharacteristics, and any fail messages
-using TestResult = std::pair<const FanCharacteristics, const FailState>;
-using enable_mode_t = int;
-using slope_t = double;
-
-enum class FanState { unknown, stopped, full_speed };
+const Pwm pwm_min_abs = 0, pwm_max_abs = 255;
 
 class FanInterface {
 public:
-  FanInterface(const UID &uid, const fan::Config &conf, bool dynamic,
-               enable_mode_t driverEnableMode,
-               enable_mode_t manualEnableMode = 1);
+  FanInterface() = default;
+  explicit FanInterface(string label_);
   virtual ~FanInterface() = default;
 
-  vector<fan::Point> points;
-  decltype(points)::iterator
-  prev_it;
+  string label;
+  shared_ptr<fc::SensorInterface> sensor;
+  Rpm_to_Pwm_Map rpm_to_pwm;
+  Temp_to_Rpm_Map temp_to_rpm;
+  Pwm start_pwm = 0;
+  milliseconds interval{0};
+  bool ignore{false};
 
-  pwm_t pwm_start;
-  milliseconds wait_time;
+  bool update();
+  virtual void test(shared_ptr<double> completed);
+  bool tested() const;
+  bool configured() const;
 
-  const enable_mode_t manual_enable_mode;
-  enable_mode_t driver_enable_mode;
+  virtual bool enable_control() const = 0;
+  virtual bool disable_control() const = 0;
 
-  // TODO Functions don't need to be public
-  virtual rpm_t readRPM() = 0;
-  virtual rpm_t readPWM() = 0;
-  virtual bool writePWM(const pwm_t &pwm) = 0;
-  virtual bool writeEnableMode(const enable_mode_t &mode) = 0;
-  bool recoverControl(const string &deviceLabel);
+  virtual void from(const fc_pb::Fan &f, const SensorMap &sensor_map);
+  virtual void to(fc_pb::Fan &f) const = 0;
+  virtual bool valid() const = 0;
+  virtual string_view uid() const = 0;
 
-  void update(const temp_t temp);
-
-  TestResult test();
-  static void writeTestResult(const UID &uid, const FanCharacteristics &result,
-                              DeviceType devType);
+  friend std::ostream &operator<<(std::ostream &os, const FanInterface &f);
 
 protected:
-  const hwid_t hw_id;
-  const string hw_id_str;
+  struct {
+    bool just_started{true};
+    int rem_intervals{0};
+    Rpm targeted_rpm{0};
+    int top_stickiness_rem_intervals{0};
+  } smoothing;
 
-  const bool dynamic;
+  virtual bool set_pwm(const Pwm pwm) const = 0;
+  virtual Pwm get_pwm() const = 0;
+  bool set_rpm(Rpm rpm);
+  virtual Rpm get_rpm() const = 0;
+  bool recover_control() const;
+  Rpm smooth_rpm(const Rpm rpm);
+  void sleep_for_interval() const;
 
-  pwm_t calcPWM(const rpm_t &rpmMin, const pwm_t &pwmMin,
-                const slope_t &slope, const rpm_t &rpm);
-  void validatePoints(const UID &fanUID, const bool &tested,
-                      const FanCharacteristics &c);
+  optional<Rpm> set_stabilised_pwm(const Pwm pwm) const;
+  bool set_pwm_test() const;
+  void test_stopped(Pwm_to_Rpm_Map &pwm_to_rpm);
+  void test_start(Pwm_to_Rpm_Map &pwm_to_rpm);
+  void test_interval(Pwm_to_Rpm_Map &pwm_to_rpm);
+  void test_running_min(Pwm_to_Rpm_Map &pwm_to_rpm);
+  void test_mapping(Pwm_to_Rpm_Map &pwm_to_rpm);
 
-  pwm_t testPWM(const FanCharacteristics &c, const rpm_t &rpm); // TODO remove
-
-  rpm_t getMaxRPM(FanState &state, const milliseconds &waitTime);
-  milliseconds getMaxSpeedChangeTime(FanState &state, const rpm_t &rpmMax);
-  pwm_t getMaxPWM(FanState &state, const milliseconds &waitTime,
-                  const rpm_t &rpmMax);
-  pwm_t getPWMStart(FanState &state, const milliseconds &waitTime);
-  tuple<pwm_t, rpm_t> getMinAttributes(FanState &state,
-                                       const milliseconds &waitTime,
-                                       const pwm_t &startPWM);
+  Percent rpm_to_percent(const Rpm rpm) const;
+  Rpm percent_to_rpm(Percent percent) const;
+  Rpm pwm_to_rpm(Pwm pwm) const;
+  void temp_to_rpm_from(const string &src);
+  void rpm_to_pwm_from(const string &src);
+  void rpm_to_pwm_from(const Pwm_to_Rpm_Map &pwm_to_rpm);
 };
 
-struct FanCharacteristicPaths {
-  explicit FanCharacteristicPaths(const UID &uid);
-
-  DeviceType type;
-
-  constexpr static const char *pwm_prefix = "pwm", *rpm_prefix = "fan";
-  string pwm_min_pf, pwm_max_pf, pwm_start_pf,
-      rpm_min_pf, rpm_max_pf, slope_pf, wait_time_pf;
-  const string hw_id;
-
-  bool tested() const;
-  static int getDeviceID(const UID &uid) { return Util::lastNum(uid.dev_name); }
-};
-
-struct FanCharacteristics {
-  FanCharacteristics() = default;
-  FanCharacteristics(rpm_t rpmMin, rpm_t rpmMax, pwm_t pwmStart, pwm_t pwmMin,
-                     pwm_t pwmMax, slope_t slope, milliseconds waitTime)
-      : rpm_min(rpmMin), rpm_max(rpmMax), pwm_start(pwmStart),
-        pwm_min(pwmMin), pwm_max(pwmMax), slope(slope), wait_time(waitTime) {}
-
-  rpm_t rpm_min, rpm_max;
-  pwm_t pwm_start, pwm_min{fancon::fan::pwm_min_abs}, pwm_max;
-  slope_t slope;
-  milliseconds wait_time;
-
-  bool valid() const {
-    return (rpm_min > 0) && (rpm_min < rpm_max) && (pwm_min > 0) &&
-        (pwm_max <= 255) && (pwm_min < pwm_max) && (pwm_start > 0) &&
-        (slope > 0);
-  }
-};
-}
+std::ostream &operator<<(std::ostream &os, const fc::FanInterface &f);
+} // namespace fc
 
 #endif // FANCON_FANINTERFACE_HPP
