@@ -1,35 +1,51 @@
 #ifndef FANCON_SENSORINTERFACE_HPP
 #define FANCON_SENSORINTERFACE_HPP
 
-#include "Config.hpp"
-#include "UID.hpp"
+#include <mutex>
+#include <numeric>
 
-using fancon::UID;
-using fancon::temp_t;
+#include "Util.hpp"
+#include "proto/DevicesSpec.pb.h"
 
-namespace fancon {
-struct SensorInterface {
-  explicit SensorInterface(temp_t temp = {}, bool update = true)
-      : temp(temp), update(update) {}
+using Temp = int;
 
-  temp_t temp;
-  bool update;
+namespace fc {
+extern uint temp_averaging_intervals;
 
-  virtual bool operator==(const UID &other) const = 0;
+class SensorInterface {
+public:
+  SensorInterface() = default;
+  explicit SensorInterface(string label_);
+  virtual ~SensorInterface() = default;
 
-  virtual temp_t read() const = 0;
-  void refresh();
+  string label;
+  bool ignore{false};
+
+  Temp get_average_temp();
+  virtual optional<Temp> min_temp() const { return nullopt; }
+  virtual optional<Temp> max_temp() const { return nullopt; }
+
+  virtual void from(const fc_pb::Sensor &s);
+  virtual void to(fc_pb::Sensor &s) const;
+  virtual bool valid() const = 0;
+  virtual string_view uid() const = 0;
+
+  friend std::ostream &operator<<(std::ostream &os, const SensorInterface &s);
+
+protected:
+  std::mutex read_mutex;
+  chrono::high_resolution_clock::time_point last_read_time;
+  vector<Temp> temp_history;
+  size_t temp_history_i = 0;
+  Temp last_avg_temp = 0;
+
+  virtual Temp read() const = 0;
+  bool fresh() const;
 };
 
-struct Sensor : public SensorInterface {
-  explicit Sensor(const UID &uid);
+std::ostream &operator<<(std::ostream &os, const SensorInterface &s);
+} // namespace fc
 
-  string input_path;
-
-  bool operator==(const UID &other) const override;
-
-  temp_t read() const override { return Util::read<temp_t>(input_path) / 1000; }
-};
-}
+using SensorMap = std::unordered_map<string, shared_ptr<fc::SensorInterface>>;
 
 #endif // FANCON_SENSORINTERFACE_HPP
